okay ... vhost :)

+++ commands/vhost.py
@@ -21,14 +21,21 @@ class vhost(pyserv.Command):
 			elif len(arg[0]) < 6:
 				self.msg(source, "Your vhost is too short.")
 			else:
-				self.query("delete from vhosts where user = '%s'" % self.auth(source))
-				self.query("insert into vhosts values ('%s','%s','0')" % (self.auth(source), arg[0]))
-				self.msg(source, "Your new vhost\2 %s\2 has been requested" % arg[0])
-				for data in self.query("select host,username from online where uid = '%s'" % source):
-					self.send(":%s CHGIDENT %s %s" % (self.bot, source, data[1]))
-					self.send(":%s CHGHOST %s %s" % (self.bot, source, data[0]))
-				for data in self.query("select uid from opers"):
-					self.msg(data[0], "vHost request received from\2 %s\2" % self.auth(source))
+				entry = False
+				for data in self.query("select user from vhosts where vhost = '%s' and user != '%s'" % (arg[0], self.auth(source))):
+					user = data[0]
+					entry = True
+				if not entry:
+					self.query("delete from vhosts where user = '%s'" % self.auth(source))
+					self.query("insert into vhosts values ('%s','%s','0')" % (self.auth(source), arg[0]))
+					self.msg(source, "Your new vhost\2 %s\2 has been requested" % arg[0])
+					for data in self.query("select host,username from online where uid = '%s'" % source):
+						self.send(":%s CHGIDENT %s %s" % (self.bot, source, data[1]))
+						self.send(":%s CHGHOST %s %s" % (self.bot, source, data[0]))
+					for data in self.query("select uid from opers"):
+						self.msg(data[0], "vHost request received from\2 %s\2" % self.auth(source))
+				else:
+					self.msg(source, "%s is already using this vHost." % user)
 		elif len(arg) == 0:
 			self.query("delete from vhosts where user = '%s'" % self.auth(source))
 			self.msg(source, "Done.")
